requirements.txt
streamlit
openai
python-dotenv
SpeechRecognition
pyttsx3
pyaudio